feat: create new main.go entry point
Create a new simple main.go entry point that calls architect.Main(). This is the final refactoring step for the main.go file. The existing main.go has been moved to legacy_main.go with appropriate build tags to maintain testing compatibility.

- Created a new root-level main.go that simply calls architect.Main()
- Moved all transitional code to legacy_main.go with build tags
- Fixed various testing issues to ensure backward compatibility
- Fixed Go vet errors including adding missing methods in mock loggers
- Updated tokenInfo result handling in getTokenInfo
- Updated GetErrorDetails to handle nil errors
