+++ cmd/architect/prompt.go
