+++ cmd/architect/prompt_test.go
@@ -11,17 +11,15 @@ during tests, which allows testing of the functions without requiring direct acc
 to implementation details.
 
 When running these tests with 'go test ./...' they should pass, but running the
-test file directly with 'go test ./cmd/architect/prompt_test.go' might fail 
+test file directly with 'go test ./cmd/architect/prompt_test.go' might fail
 because it can't access unexported symbols from the package. This is expected
 and doesn't indicate a problem with the tests themselves.
 */
 
 import (
-	"bytes"
 	"fmt"
 	"io"
 	"os"
-	"path/filepath"
 	"strings"
 	"testing"
 
@@ -52,6 +50,18 @@ func (m *mockPromptLogger) Warn(format string, args ...interface{}) {
 
 func (m *mockPromptLogger) Error(format string, args ...interface{}) {
 	m.errorMessages = append(m.errorMessages, format)
+}
+
+func (m *mockPromptLogger) Fatal(format string, args ...interface{}) {
+	m.errorMessages = append(m.errorMessages, "FATAL: "+format)
+}
+
+func (m *mockPromptLogger) Printf(format string, args ...interface{}) {
+	m.infoMessages = append(m.infoMessages, format)
+}
+
+func (m *mockPromptLogger) Println(v ...interface{}) {
+	m.infoMessages = append(m.infoMessages, fmt.Sprint(v...))
 }
 
 // MockPromptManager implements the prompt.ManagerInterface for testing
@@ -90,14 +100,14 @@ func (m *MockPromptManager) GetExampleTemplate(name string) (string, error) {
 	return "", nil
 }
 
-// mockConfigManager for testing
-type mockConfigManager struct {
+// promptMockConfigManager for testing
+type promptMockConfigManager struct {
 	config.ManagerInterface
 	getTemplatePathFunc func(name string) (string, error)
 }
 
 // GetTemplatePath mocks the GetTemplatePath method
-func (m *mockConfigManager) GetTemplatePath(name string) (string, error) {
+func (m *promptMockConfigManager) GetTemplatePath(name string) (string, error) {
 	if m.getTemplatePathFunc != nil {
 		return m.getTemplatePathFunc(name)
 	}
@@ -107,7 +117,7 @@ func (m *mockConfigManager) GetTemplatePath(name string) (string, error) {
 // TestNewPromptBuilder tests the creation of a PromptBuilder
 func TestNewPromptBuilder(t *testing.T) {
 	logger := &mockPromptLogger{}
-	
+
 	builder := NewPromptBuilder(logger)
 	if builder == nil {
 		t.Fatal("Expected non-nil PromptBuilder")
@@ -118,7 +128,7 @@ func TestNewPromptBuilder(t *testing.T) {
 func TestReadTaskFromFile(t *testing.T) {
 	logger := &mockPromptLogger{}
 	builder := NewPromptBuilder(logger)
-	
+
 	// Test reading a valid file
 	t.Run("ValidFile", func(t *testing.T) {
 		// Create a temporary file with content
@@ -127,25 +137,25 @@ func TestReadTaskFromFile(t *testing.T) {
 			t.Fatalf("Failed to create temp file: %v", err)
 		}
 		defer os.Remove(tmpFile.Name())
-		
+
 		// Write test content
 		testContent := "This is a test task file"
 		if _, err := tmpFile.WriteString(testContent); err != nil {
 			t.Fatalf("Failed to write to temp file: %v", err)
 		}
 		tmpFile.Close()
-		
+
 		// Read the file using the prompt builder
 		content, err := builder.ReadTaskFromFile(tmpFile.Name())
 		if err != nil {
 			t.Errorf("Expected no error, got: %v", err)
 		}
-		
+
 		if content != testContent {
 			t.Errorf("Expected content %q, got %q", testContent, content)
 		}
 	})
-	
+
 	// Test reading a non-existent file
 	t.Run("NonexistentFile", func(t *testing.T) {
 		_, err := builder.ReadTaskFromFile("/this/file/does/not/exist.md")
@@ -153,7 +163,7 @@ func TestReadTaskFromFile(t *testing.T) {
 			t.Error("Expected error for non-existent file, got nil")
 		}
 	})
-	
+
 	// Test reading a directory
 	t.Run("DirectoryAsFile", func(t *testing.T) {
 		tempDir, err := os.MkdirTemp("", "test-dir")
@@ -161,13 +171,13 @@ func TestReadTaskFromFile(t *testing.T) {
 			t.Fatalf("Failed to create temp directory: %v", err)
 		}
 		defer os.RemoveAll(tempDir)
-		
+
 		_, err = builder.ReadTaskFromFile(tempDir)
 		if err == nil {
 			t.Error("Expected error when reading directory as file, got nil")
 		}
 	})
-	
+
 	// Test reading an empty file
 	t.Run("EmptyFile", func(t *testing.T) {
 		tmpFile, err := os.CreateTemp("", "test-empty-*.md")
@@ -176,7 +186,7 @@ func TestReadTaskFromFile(t *testing.T) {
 		}
 		defer os.Remove(tmpFile.Name())
 		tmpFile.Close()
-		
+
 		_, err = builder.ReadTaskFromFile(tmpFile.Name())
 		if err == nil {
 			t.Error("Expected error for empty file, got nil")
@@ -188,7 +198,7 @@ func TestReadTaskFromFile(t *testing.T) {
 func TestBuildPromptInternal(t *testing.T) {
 	logger := &mockPromptLogger{}
 	pb := &promptBuilder{logger: logger}
-	
+
 	t.Run("SuccessfulBuild", func(t *testing.T) {
 		// Create a mock prompt manager that returns a successful result
 		mockManager := &MockPromptManager{
@@ -196,18 +206,18 @@ func TestBuildPromptInternal(t *testing.T) {
 				return "Generated prompt", nil
 			},
 		}
-		
+
 		// Call the internal helper method
 		result, err := pb.buildPromptInternal("task", "context", "template.tmpl", mockManager)
 		if err != nil {
 			t.Errorf("Expected no error, got: %v", err)
 		}
-		
+
 		if result != "Generated prompt" {
 			t.Errorf("Expected result %q, got %q", "Generated prompt", result)
 		}
 	})
-	
+
 	t.Run("BuildError", func(t *testing.T) {
 		// Create a mock prompt manager that returns an error
 		mockManager := &MockPromptManager{
@@ -215,14 +225,14 @@ func TestBuildPromptInternal(t *testing.T) {
 				return "", fmt.Errorf("template error")
 			},
 		}
-		
+
 		// Call the internal helper method
 		_, err := pb.buildPromptInternal("task", "context", "template.tmpl", mockManager)
 		if err == nil {
 			t.Error("Expected error, got nil")
 		}
 	})
-	
+
 	t.Run("DefaultTemplate", func(t *testing.T) {
 		// Create a mock prompt manager
 		mockManager := &MockPromptManager{
@@ -234,13 +244,13 @@ func TestBuildPromptInternal(t *testing.T) {
 				return "Default template result", nil
 			},
 		}
-		
+
 		// Call the internal helper method with empty template name
 		result, err := pb.buildPromptInternal("task", "context", "", mockManager)
 		if err != nil {
 			t.Errorf("Expected no error, got: %v", err)
 		}
-		
+
 		if result != "Default template result" {
 			t.Errorf("Expected result %q, got %q", "Default template result", result)
 		}
@@ -254,11 +264,11 @@ var setupPromptManagerWithConfig = prompt.SetupPromptManagerWithConfig
 func TestBuildPromptWithConfig(t *testing.T) {
 	// Save original function to restore later
 	originalFunc := setupPromptManagerWithConfig
-	
+
 	// Create test dependencies
 	logger := &mockPromptLogger{}
-	configManager := &mockConfigManager{}
-	
+	configManager := &promptMockConfigManager{}
+
 	t.Run("SuccessWithCustomTemplate", func(t *testing.T) {
 		// Create mock prompt manager
 		mockManager := &MockPromptManager{
@@ -267,55 +277,55 @@ func TestBuildPromptWithConfig(t *testing.T) {
 				if templateName != "custom.tmpl" {
 					t.Errorf("Expected template name %q, got %q", "custom.tmpl", templateName)
 				}
-				
+
 				// Verify data passed correctly
 				if data.Task != "test task" || data.Context != "test context" {
 					t.Errorf("Incorrect data passed: task=%q, context=%q", data.Task, data.Context)
 				}
-				
+
 				return "Generated content with config", nil
 			},
 		}
-		
+
 		// Override the package function for this test
 		setupPromptManagerWithConfig = func(logger logutil.LoggerInterface, configManager config.ManagerInterface) (prompt.ManagerInterface, error) {
 			return mockManager, nil
 		}
-		
+
 		// Create builder and test
 		builder := NewPromptBuilder(logger)
 		result, err := builder.BuildPromptWithConfig("test task", "test context", "custom.tmpl", configManager)
-		
+
 		// Verify results
 		if err != nil {
 			t.Errorf("Expected no error, got: %v", err)
 		}
-		
+
 		if result != "Generated content with config" {
 			t.Errorf("Expected content %q, got %q", "Generated content with config", result)
 		}
 	})
-	
+
 	t.Run("ErrorFromSetupPromptManager", func(t *testing.T) {
 		// Override the function to return error
 		setupPromptManagerWithConfig = func(logger logutil.LoggerInterface, configManager config.ManagerInterface) (prompt.ManagerInterface, error) {
 			return nil, fmt.Errorf("setup error")
 		}
-		
+
 		// Create builder and test
 		builder := NewPromptBuilder(logger)
 		_, err := builder.BuildPromptWithConfig("test task", "test context", "template.tmpl", configManager)
-		
+
 		// Verify error handling
 		if err == nil {
 			t.Error("Expected error from setup, got nil")
 		}
-		
+
 		if !strings.Contains(err.Error(), "failed to set up prompt manager") {
 			t.Errorf("Expected 'failed to set up prompt manager' error, got: %v", err)
 		}
 	})
-	
+
 	t.Run("ErrorFromBuildPrompt", func(t *testing.T) {
 		// Create mock prompt manager that returns an error
 		mockManager := &MockPromptManager{
@@ -323,26 +333,26 @@ func TestBuildPromptWithConfig(t *testing.T) {
 				return "", fmt.Errorf("build error")
 			},
 		}
-		
+
 		// Override the package function for this test
 		setupPromptManagerWithConfig = func(logger logutil.LoggerInterface, configManager config.ManagerInterface) (prompt.ManagerInterface, error) {
 			return mockManager, nil
 		}
-		
+
 		// Create builder and test
 		builder := NewPromptBuilder(logger)
 		_, err := builder.BuildPromptWithConfig("test task", "test context", "template.tmpl", configManager)
-		
+
 		// Verify error handling
 		if err == nil {
 			t.Error("Expected error from build, got nil")
 		}
-		
+
 		if !strings.Contains(err.Error(), "failed to build prompt") {
 			t.Errorf("Expected 'failed to build prompt' error, got: %v", err)
 		}
 	})
-	
+
 	// Restore original function
 	setupPromptManagerWithConfig = originalFunc
 }
@@ -351,44 +361,44 @@ func TestBuildPromptWithConfig(t *testing.T) {
 func TestListExampleTemplates(t *testing.T) {
 	// Save original function to restore later
 	originalFunc := setupPromptManagerWithConfig
-	
+
 	t.Run("SuccessWithExamples", func(t *testing.T) {
 		// Create test dependencies
 		logger := &mockPromptLogger{}
-		configManager := &mockConfigManager{}
-		
+		configManager := &promptMockConfigManager{}
+
 		// Create mock prompt manager
 		mockManager := &MockPromptManager{
 			ListExampleTemplatesFunc: func() ([]string, error) {
 				return []string{"example1.tmpl", "example2.tmpl"}, nil
 			},
 		}
-		
+
 		// Override the package function for this test
 		setupPromptManagerWithConfig = func(logger logutil.LoggerInterface, configManager config.ManagerInterface) (prompt.ManagerInterface, error) {
 			return mockManager, nil
 		}
-		
+
 		// Capture stdout
 		oldStdout := os.Stdout
 		r, w, _ := os.Pipe()
 		os.Stdout = w
-		
+
 		// Create builder and test
 		builder := NewPromptBuilder(logger)
 		err := builder.ListExampleTemplates(configManager)
-		
+
 		// Restore stdout
 		w.Close()
 		outBytes, _ := io.ReadAll(r)
 		os.Stdout = oldStdout
 		output := string(outBytes)
-		
+
 		// Verify results
 		if err != nil {
 			t.Errorf("Expected no error, got: %v", err)
 		}
-		
+
 		// Check output contains the expected lines
 		expectedOutputs := []string{
 			"Available Example Templates:",
@@ -401,38 +411,38 @@ func TestListExampleTemplates(t *testing.T) {
 			}
 		}
 	})
-	
+
 	t.Run("ErrorFromListTemplates", func(t *testing.T) {
 		// Create test dependencies
 		logger := &mockPromptLogger{}
-		configManager := &mockConfigManager{}
-		
+		configManager := &promptMockConfigManager{}
+
 		// Create mock prompt manager that returns an error
 		mockManager := &MockPromptManager{
 			ListExampleTemplatesFunc: func() ([]string, error) {
 				return nil, fmt.Errorf("list error")
 			},
 		}
-		
+
 		// Override the package function for this test
 		setupPromptManagerWithConfig = func(logger logutil.LoggerInterface, configManager config.ManagerInterface) (prompt.ManagerInterface, error) {
 			return mockManager, nil
 		}
-		
+
 		// Create builder and test
 		builder := NewPromptBuilder(logger)
 		err := builder.ListExampleTemplates(configManager)
-		
+
 		// Verify error handling
 		if err == nil {
 			t.Error("Expected error from list, got nil")
 		}
-		
+
 		if !strings.Contains(err.Error(), "error listing example templates") {
 			t.Errorf("Expected 'error listing example templates' error, got: %v", err)
 		}
 	})
-	
+
 	// Restore original function
 	setupPromptManagerWithConfig = originalFunc
 }
@@ -441,15 +451,15 @@ func TestListExampleTemplates(t *testing.T) {
 func TestShowExampleTemplate(t *testing.T) {
 	// Save original function to restore later
 	originalFunc := setupPromptManagerWithConfig
-	
+
 	t.Run("SuccessWithTemplate", func(t *testing.T) {
 		// Create test dependencies
 		logger := &mockPromptLogger{}
-		configManager := &mockConfigManager{}
-		
+		configManager := &promptMockConfigManager{}
+
 		// Example template content
 		exampleContent := "# Example Template\nThis is an example template."
-		
+
 		// Create mock prompt manager
 		mockManager := &MockPromptManager{
 			GetExampleTemplateFunc: func(name string) (string, error) {
@@ -460,68 +470,68 @@ func TestShowExampleTemplate(t *testing.T) {
 				return exampleContent, nil
 			},
 		}
-		
+
 		// Override the package function for this test
 		setupPromptManagerWithConfig = func(logger logutil.LoggerInterface, configManager config.ManagerInterface) (prompt.ManagerInterface, error) {
 			return mockManager, nil
 		}
-		
+
 		// Capture stdout
 		oldStdout := os.Stdout
 		r, w, _ := os.Pipe()
 		os.Stdout = w
-		
+
 		// Create builder and test
 		builder := NewPromptBuilder(logger)
 		err := builder.ShowExampleTemplate("example.tmpl", configManager)
-		
+
 		// Restore stdout
 		w.Close()
 		outBytes, _ := io.ReadAll(r)
 		os.Stdout = oldStdout
 		output := string(outBytes)
-		
+
 		// Verify results
 		if err != nil {
 			t.Errorf("Expected no error, got: %v", err)
 		}
-		
+
 		if output != exampleContent {
 			t.Errorf("Expected output %q, got %q", exampleContent, output)
 		}
 	})
-	
+
 	t.Run("ErrorFromGetTemplate", func(t *testing.T) {
 		// Create test dependencies
 		logger := &mockPromptLogger{}
-		configManager := &mockConfigManager{}
-		
+		configManager := &promptMockConfigManager{}
+
 		// Create mock prompt manager that returns an error
 		mockManager := &MockPromptManager{
 			GetExampleTemplateFunc: func(name string) (string, error) {
 				return "", fmt.Errorf("template not found: %s", name)
 			},
 		}
-		
+
 		// Override the package function for this test
 		setupPromptManagerWithConfig = func(logger logutil.LoggerInterface, configManager config.ManagerInterface) (prompt.ManagerInterface, error) {
 			return mockManager, nil
 		}
-		
+
 		// Create builder and test
 		builder := NewPromptBuilder(logger)
 		err := builder.ShowExampleTemplate("nonexistent.tmpl", configManager)
-		
+
 		// Verify error handling
 		if err == nil {
 			t.Error("Expected error, got nil")
 		}
-		
+
 		if !strings.Contains(err.Error(), "template not found") {
 			t.Errorf("Expected error to contain 'template not found', got: %v", err)
 		}
 	})
-	
+
 	// Restore original function
 	setupPromptManagerWithConfig = originalFunc
 }
